Rename and restructure CMake (#1)
# Changes

The `install(EXPORT ...)` commands were failing because they required
Boost to also `install(EXPORT ...)` which Boost doesn't when you
submodule the repository. As such I've made it so that only if you load
the repository as the root repository will it setup the exporting.

As I was resolving this issue, I've also went ahead and renamed the
original targets because it was too generic and also changed the
namespace to be similar to Boost. Yes its not part of the Boost library
but when your `#include` starts with `boost` I think there are some
overlap.